+++ apps/myaccount/src/api/multi-factor-push.ts
@@ -18,7 +18,6 @@
 
 import { AsgardeoSPAClient, HttpClientInstance, HttpRequestConfig } from "@asgardeo/auth-react";
 import { AxiosError, AxiosResponse } from "axios";
-import { HttpErrorResponseDataInterface } from "@wso2is/core/models";
 import { HttpMethods } from "../models";
 import { store } from "../store";
 
@@ -48,7 +47,7 @@ export const initPushAuthenticatorQRCode = (): Promise<any> => {
 
             return Promise.resolve(response);
         })
-        .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        .catch((error: AxiosError) => {
             return Promise.reject(error);
         });
 };
@@ -70,7 +69,7 @@ export const getPushEnabledDevices = () => {
 
             return Promise.resolve(response);
         })
-        .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        .catch((error: AxiosError) => {
             return Promise.reject(error);
         });
 };
@@ -95,7 +94,7 @@ export const deletePushAuthRegisteredDevice = (deviceId: string) => {
 
             return Promise.resolve(response);
         })
-        .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        .catch((error: AxiosError) => {
             return Promise.reject(error);
         });
 };