+++ apps/myaccount/src/components/account-recovery/options/sms-recovery/sms-recovery.tsx
@@ -16,9 +16,7 @@
  * under the License.
  */
 
-import { IdentifiableComponentInterface,
-    HttpErrorResponseDataInterface
-} from "@wso2is/core/models";
+import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import { FinalForm, FinalFormField, FormRenderProps, FormValue, TextFieldAdapter } from "@wso2is/form";
 import { Button, GenericIcon, Hint, PrimaryButton, SecondaryButton } from "@wso2is/react-components";
 import { FormValidation } from "@wso2is/validation";
@@ -155,7 +153,7 @@ export const SMSRecovery: React.FunctionComponent<SMSRecoveryProps> = (
                 dispatch(getProfileInformation());
                 dispatch(setActiveForm(null));
             })
-            .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+            .catch((error: AxiosError) => {
                 if (error?.response?.data?.detail) {
                     onAlertFired({
                         description: t(