+++ features/admin.ask-password-flow-builder.v1/providers/ask-password-flow-builder-provider.tsx
@@ -36,9 +36,7 @@ import {
     UpdateGovernanceConnectorConfigInterface
 } from "@wso2is/admin.server-configurations.v1/models/governance-connectors";
 import { GovernanceConnectorUtils } from "@wso2is/admin.server-configurations.v1/utils/governance-connector-utils";
-import { AlertLevels,
-    HttpErrorResponseDataInterface
-} from "@wso2is/core/models";
+import { AlertLevels } from "@wso2is/core/models";
 import { addAlert } from "@wso2is/core/store";
 import { useReactFlow } from "@xyflow/react";
 import { AxiosError } from "axios";
@@ -132,7 +130,7 @@ const FlowContextWrapper: FC<AskPasswordFlowBuilderProviderProps> = ({
     *
     * @param error - Axios error object.
     */
-    const handleUpdateError = (error: AxiosError<HttpErrorResponseDataInterface>) => {
+    const handleUpdateError = (error: AxiosError) => {
         if (error.response && error.response.data && error.response.data.detail) {
             dispatch(
                 addAlert({
@@ -237,7 +235,7 @@ const FlowContextWrapper: FC<AskPasswordFlowBuilderProviderProps> = ({
             .then(() => {
                 handleUpdateSuccess();
             })
-            .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+            .catch((error: AxiosError) => {
                 handleUpdateError(error);
             }).finally(() => {
                 // Reset the updated state.