+++ features/admin.identity-providers.v1/api/fido-configs.ts
@@ -26,9 +26,7 @@ import useRequest, {
     RequestResultInterface
 } from "@wso2is/admin.core.v1/hooks/use-request";
 import { IdentityAppsApiException } from "@wso2is/core/exceptions";
-import { HttpMethods,
-    HttpErrorResponseDataInterface
-} from "@wso2is/core/models";
+import { HttpMethods } from "@wso2is/core/models";
 import { AxiosError, AxiosResponse } from "axios";
 import { FIDOConfigsInterface, FIDOConnectorConfigsInterface } from "../models";
 
@@ -105,7 +103,7 @@ export const updateFidoConfigs = (
             }
 
             return Promise.resolve(response.data as FIDOConnectorConfigsInterface);
-        }).catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        }).catch((error: AxiosError) => {
             throw new IdentityAppsApiException(
                 ConnectionUIConstants.ERROR_MESSAGES.FIDO_AUTHENTICATOR_CONFIG_UPDATE_ERROR,
                 error?.stack,