+++ features/admin.organization-discovery.v1/api/update-organization-discovery-config.ts
@@ -18,9 +18,7 @@
 
 import { AsgardeoSPAClient, HttpClientInstance } from "@asgardeo/auth-react";
 import { store } from "@wso2is/admin.core.v1/store";
-import { HttpMethods,
-    HttpErrorResponseDataInterface
-} from "@wso2is/core/models";
+import { HttpMethods } from "@wso2is/core/models";
 import { AxiosError, AxiosRequestConfig, AxiosResponse } from "axios";
 import { OrganizationDiscoveryConfigInterface } from "../models/organization-discovery";
 
@@ -54,7 +52,7 @@ export const updateOrganizationDiscoveryConfig = (
             }
 
             return Promise.resolve(response?.data);
-        }).catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        }).catch((error: AxiosError) => {
             return Promise.reject(error?.response?.data);
         });
 };