+++ features/admin.roles.v2/components/application-roles.tsx
@@ -33,9 +33,7 @@ import { history } from "@wso2is/admin.core.v1/helpers/history";
 import { RequestErrorInterface } from "@wso2is/admin.core.v1/hooks/use-request";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models/config";
 import { AppState } from "@wso2is/admin.core.v1/store";
-import { AlertLevels, IdentifiableComponentInterface,
-    HttpErrorResponseDataInterface
-} from "@wso2is/core/models";
+import { AlertLevels, IdentifiableComponentInterface } from "@wso2is/core/models";
 import { addAlert } from "@wso2is/core/store";
 import {
     ConfirmationModal,
@@ -257,7 +255,7 @@ export const ApplicationRoles: FunctionComponent<ApplicationRolesSettingsInterfa
                     if (shouldUpdateSelectedRolesList) {
                         setSelectedRoles(rolesArray);
                     }
-                }).catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+                }).catch((error: AxiosError) => {
                     if (error?.response?.data?.description) {
                         dispatch(addAlert({
                             description: error?.response?.data?.description ??
@@ -322,7 +320,7 @@ export const ApplicationRoles: FunctionComponent<ApplicationRolesSettingsInterfa
 
                 onUpdate(appId);
             })
-            .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+            .catch((error: AxiosError) => {
                 if (error.response && error.response.data && error.response.data.description) {
                     dispatch(addAlert({
                         description: error.response.data.description,