+++ features/admin.users.v1/api/profile.ts
@@ -30,8 +30,7 @@ import {
     HttpMethods,
     ProfileInfoInterface,
     ProfileSchemaInterface,
-    SchemaResponseInterface,
-    HttpErrorResponseDataInterface
+    SchemaResponseInterface
 } from "@wso2is/core/models";
 import { AxiosError, AxiosResponse } from "axios";
 import { AttributeDataType } from "../constants";
@@ -121,7 +120,7 @@ export const getProfileInfo = (endpoint: string,
 
             return Promise.resolve(profileResponse);
         })
-        .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        .catch((error: AxiosError) => {
             // Check if the API responds with a `500` error, if it does,
             // navigate the user to the login error page.
             if (error.response
@@ -177,7 +176,7 @@ export const updateProfileInfo = (info: Record<string, unknown>): Promise<Profil
 
             return Promise.resolve(response.data);
         })
-        .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        .catch((error: AxiosError) => {
             throw new IdentityAppsApiException(
                 ProfileConstants.PROFILE_INFO_UPDATE_REQUEST_ERROR,
                 error.stack,
@@ -278,7 +277,7 @@ export const getProfileSchemas = (): Promise<ProfileSchemaInterface[]> => {
 
             return Promise.resolve(schemaAttributes as ProfileSchemaInterface[]);
         })
-        .catch((error: AxiosError<HttpErrorResponseDataInterface>) => {
+        .catch((error: AxiosError) => {
             throw new IdentityAppsApiException(
                 ProfileConstants.SCHEMA_FETCH_REQUEST_ERROR,
                 error.stack,